cybersec-web/00-submit/requirements.txt
beautifulsoup4==4.12.3
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
colorama==0.4.6
idna==3.10
lxml==5.3.0
pillow==11.2.1
requests==2.32.3
soupsieve==2.6
urllib3==2.2.3
